backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
numpy>=2.0.0,<3.0.0
scikit-learn==1.5.0
opencv-python-headless>=4.10.0
mediapipe>=0.10.8
librosa==0.10.1
soundfile==0.12.1
Pillow>=10.1.0
PyJWT==2.8.0
werkzeug==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
eventlet==0.35.1
requests==2.31.0
huggingface-hub==0.34.4
